Vagrantfile: Update Vagrant box
The boxcutter box was no longer available. I tried the official Ubuntu
Xenial 16.04 daily build box: in the past there were issues getting it
to work. Nowadays it works, but it seems to write a logfile to the
project root and I could find no way to prevent it from doing this, so
for now the geerlingguy box seems like the best available option.