Remove "pageelements" templates
These were intended to be a bunch of pre-packaged partials to use with a "Page Elements" matrix field. However, this seems too opinionated to include in the default templates: it's not necessarily obvious how you should use them, and in practice it probably isn't possible to make this sort of thing "one size fits all". Consequently, I'm opting to remove them for simplicity's sake.