Package.json: update all dependencies to more recent versions
These were updated automatically by running `ncu`[1]

In addition to this, I had to add `babel-core` as a dependency, since it
seems it's required by the latest version of `babelify`.

Having also updated to a newer version of npm, this also created a
`package-lock.json` file, which did not exist with earlier versions of
npm.

[1] https://www.npmjs.com/package/npm-check-updates